requirements.txt
nltk
transformers
torch
google-genai
pytest
